Make a module so we can see the API
